refactor: :recycle: Move makeContentServerSchema to ~/schemas/serverSchemas/contents.
