+++ schemas/serverSchemas/index.ts
@@ -1,3 +1,4 @@
 export * from "./articles";
 export * from "./files";
 export * from "./users";
+export * from "./contents";
