+++ server/utils/index.ts
@@ -7,5 +7,4 @@ export * from "./files";
 export * from "./articles";
 export * from "../../schemas/serverSchemas";
 export * from "./comments";
-export * from "./contents";
 export * from "./users";
